Relocate iw link endpoint code
